change package name to main
The name of packages which are to be compiled as modules needs to be "main" as per the official documentation (see https://golang.org/pkg/plugin/). If the package name is "main", then the compiler produces an elf shared object file. If the package name is not "main", then the compiler produces an archive file, which cannot be loaded as a plugin by the go runtime.